refactor: rmove the category folder
