+++ src/app/json/categories.json
@@ -1,7 +1,7 @@
 [
   {
     "label": "Art",
-    "link": "/categories/category/art",
+    "link": "/categories/art",
     "image": "/arts.png"
   },
   {
